New-DriveMapping and Remove-DriveMapping done

+++ Modules/Scripts/Remove-DriveMapping.ps1
@@ -26,7 +26,7 @@ param (
 	[ValidateLength(1, 1)]
 	[ValidatePattern('[A-Z]')]
 	[ValidateScript({
-		if ((Get-ItemPropertyValue "Registry::$DOSDevicesKey" -Name ('{0}:' -f $_) -ErrorAction 'SilentlyContinue') -eq $null) {
+    	if ((Get-ItemProperty 'HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\DOS Devices' | Select-Object "$_`:" -Expand "$_`:" -ErrorAction 'SilentlyContinue') -eq $null) {
 			Throw 'Drive letter not found or does not represent a mapped drive'
 		}
 		$true
@@ -41,17 +41,17 @@ param (
 
 Begin
 {
-	$DriveIconsKey = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\DriveIcons'
-	$DOSDevicesKey = 'HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\DOS Devices'
+    $DriveIconsKey = 'HKLM:\SOFTWARE\Microsoft\Windows\CurrentVersion\Explorer\DriveIcons'
+    $DOSDevicesKey = 'HKLM:\SYSTEM\CurrentControlSet\Control\Session Manager\DOS Devices'
 
-	function UnmapDriveLetter ()
+    function UnmapDriveLetter ()
 	{
 		# remove mapping
 		if ($WhatIfPreference) {
-			Write-Host "Remove-ItemProperty ""$DOSDevicesKey"" -Name ""${DriveLetter}:"" -Force" -ForegroundColor DarkYellow
+			Write-Host "Remove-ItemProperty ""$DOSDevicesKey"" -Name ""$DriveLetter`:"" -Force" -ForegroundColor DarkYellow
 		} else {
-			Write-Verbose "Remove-ItemProperty ""$DOSDevicesKey"" -Name ""${DriveLetter}:"" -Force"
-			Remove-ItemProperty "$DOSDevicesKey" -Name "${DriveLetter}:" -Force | Out-Null
+			Write-Verbose "Remove-ItemProperty ""$DOSDevicesKey"" -Name ""$DriveLetter`:"" -Force"
+			Remove-ItemProperty "$DOSDevicesKey" -Name "$DriveLetter`:" -Force | Out-Null
 		}
 		Write-Verbose 'removed persistent mapping'
 
@@ -66,36 +66,51 @@ Begin
 		}
 		Write-Verbose 'removed mapped volume label'
 
-		# remove volume label for source drive
-		if ($WhatIfPreference) {
-			Write-Host "Remove-Item ""$DriveIconsKey\$sourceLetter"" -Recurse -Force" -ForegroundColor DarkYellow
-		} else {
-			Write-Verbose "Remove-Item ""$DriveIconsKey\$sourceLetter"" -Recurse -Force"
-			if (Test-Path "$DriveIconsKey\$sourceLetter") {
-				Remove-Item "$DriveIconsKey\$sourceLetter" -Recurse -Force | Out-Null
-			}
-		}
-		Write-Verbose 'removed source drive volume label'
+        # reset source drive label only if there are no more related mappings
+        if (((Get-ItemProperty -Path $DOSDevicesKey).PSObject.Properties | ? `
+            { $_.Name -match '^[A-Z]:$' -and $_.Value -match "^\\\?\?\\$sourceLetter`:\\" }).Count -eq 0)
+        {
+		    # remove volume label for source drive
+		    if ($WhatIfPreference) {
+			    Write-Host "Remove-Item ""$DriveIconsKey\$sourceLetter"" -Recurse -Force" -ForegroundColor DarkYellow
+		    } else {
+			    Write-Verbose "Remove-Item ""$DriveIconsKey\$sourceLetter"" -Recurse -Force"
+			    if (Test-Path "$DriveIconsKey\$sourceLetter") {
+				    Remove-Item "$DriveIconsKey\$sourceLetter" -Recurse -Force | Out-Null
+			    }
+		    }
+		    Write-Verbose 'removed source drive volume label'
 
-		# restore volume label
-		if ($WhatIfPreference) {
-			Write-Host "set label of $sourceLetter drive to $label" -ForegroundColor DarkYellow
-		} else {
-			Write-Verbose "set label of $sourceLetter drive to $label"
-			Set-Volume -DriveLetter $sourceLetter -NewFileSystemLabel $label
-		}
-		Write-Verbose 'set label of source drive'
+		    # restore volume label
+		    if ($WhatIfPreference) {
+			    Write-Host "set label of $sourceLetter drive to $SourceDriveLabel" -ForegroundColor DarkYellow
+		    } else {
+			    Write-Verbose "set label of $sourceLetter drive to $SourceDriveLabel"
+			    Set-Volume -DriveLetter $sourceLetter -NewFileSystemLabel $SourceDriveLabel
+		    }
+		    Write-Verbose 'set label of source drive'
+        }
 	}
 }
 Process
 {
+    $DriveLetter = $DriveLetter.ToUpper()
+
+    $source = (Get-ItemProperty $DOSDevicesKey | Select-Object "$DriveLetter`:" -Expand "$DriveLetter`:" -ErrorAction 'SilentlyContinue')
+    if ($source -eq $null)
+    {
+        Throw '$DriveLetter is not a mapped drive'
+    }
+    # format will be something like \??\C:\foobar
+    $sourceLetter = $source[4]
+
 	if (!$SourceDriveLabel -and (Test-Path "$DriveIconsKey\$DriveLetter")) {
 		$SourceDriveLabel = (Get-ItemPropertyValue "$DriveIconsKey\$sourceLetter\DefaultLabel" -Name '(Default)' -ErrorAction 'SilentlyContinue')
 	}
 
 	if (!$SourceDriveLabel) {
 		# if no label specified then choose a default label based on the system drive or a data drive
-		$SourceDriveLabel = if ($DriveLetter -eq ($env:HOMEDRIVE)[0]) { 'System' } else { 'Data' }
+		$SourceDriveLabel = if ($sourceLetter -eq ($env:HOMEDRIVE)[0]) { 'System' } else { 'Data' }
 	}
 
 	UnmapDriveLetter
@@ -104,9 +119,9 @@ Process
 	{
 		# must reboot in order for this to take effect
 		if ($WhatIfPreference) {
-			Write-Host 'Restart-Computer -Force -Timeout 0' -ForegroundColor DarkYellow
+			Write-Host 'Restart-Computer -Force' -ForegroundColor DarkYellow
 		} else {
-			Restart-Computer -Force -Timeout 0
+			Restart-Computer -Force
 		}
 	}
 }
